Mv CollectionArtifactManifest to models.models
