+++ ansible_galaxy/build.py
@@ -10,8 +10,7 @@ import six
 from ansible_galaxy import collection_members
 from ansible_galaxy import collection_artifact_manifest
 from ansible_galaxy.collection_info import COLLECTION_INFO_FILENAME
-from ansible_galaxy.models.collection_artifact_manifest import \
-    CollectionArtifactManifest
+from ansible_galaxy.models.models import CollectionArtifactManifest
 from ansible_galaxy.utils.text import to_bytes
 
 log = logging.getLogger(__name__)
