+++ ansible_galaxy/collection_artifact_manifest.py
@@ -6,8 +6,7 @@ import os
 import yaml
 
 from ansible_galaxy.utils import chksums
-from ansible_galaxy.models.collection_artifact_manifest import \
-    CollectionArtifactManifest
+from ansible_galaxy.models.models import CollectionArtifactManifest
 from ansible_galaxy.models.models import CollectionArtifactFile
 
 log = logging.getLogger(__name__)
