+++ tests/ansible_galaxy/test_build.py
@@ -6,7 +6,7 @@ import tarfile
 from ansible_galaxy import build
 from ansible_galaxy.models.models import BuildContext
 from ansible_galaxy.models.models import CollectionInfo
-from ansible_galaxy.models.collection_artifact_manifest import CollectionArtifactManifest
+from ansible_galaxy.models.models import CollectionArtifactManifest
 
 
 log = logging.getLogger(__name__)
