+++ tests/ansible_galaxy/test_collection_artifact_manifest.py
@@ -2,7 +2,7 @@ import logging
 import os
 
 from ansible_galaxy import collection_artifact_manifest
-from ansible_galaxy.models.collection_artifact_manifest import CollectionArtifactManifest
+from ansible_galaxy.models.models import CollectionArtifactManifest
 from ansible_galaxy.models.models import CollectionArtifactFile
 from ansible_galaxy import yaml_persist
